Revert "32 bit driver support"
This reverts commit b133c554447e350fa892ba22d7a45900d7729c1d.

+++ compute-only-sample/coscommon/CosGpuCommand.h
@@ -43,7 +43,7 @@ struct GpuResourceCopy
 {
     PHYSICAL_ADDRESS    m_dstGpuAddress;
     PHYSICAL_ADDRESS    m_srcGpuAddress;
-    UINT                m_sizeBytes;
+    size_t              m_sizeBytes;
 };
 
 //
